Be more careful with poFeature being deleted.
CID 145851

I assume that the break after the stop transfer was meant to break out
of the while loop, but it only breaks out of the inner for ( iField ).
So, set poFeature to NULL when destroying it and check poFeature
before using it at the end of the while loop.


git-svn-id: https://svn.osgeo.org/gdal/trunk@30477 f0d54148-0727-0410-94bb-9a71ac55c965