Rewrite the logic in VRTSourcedRasterBand::SetMetadata.
CID 1338309 (#1 of 1): Structurally dead code (UNREACHABLE)
unreachable: Since the loop increment i++; is unreachable, the loop
body will never execute more than once

By testing if poSource is a nullptr first and returning an error if it
is a nullptr, I think the logic is much easier to follow.  I misread 
the logic here:

https://trac.osgeo.org/gdal/browser/trunk/gdal/frmts/vrt/vrtsourcedrasterband.cpp?rev=31122#L1400

and created r31355 that got it wrong.


git-svn-id: https://svn.osgeo.org/gdal/trunk@31361 f0d54148-0727-0410-94bb-9a71ac55c965