Remove CPL_IGNORE_RET_VAL on bMustComma in ogrcartotablelayer.cpp.
It was added in r32370 because bMustComma was being set to true but
never being used after the last two cases.  Removed the last two sets
and added a comment that setting to true is not currently needed in
these cases as they are not in a loop.

Complaint was:
cppcheck --inline-suppr --std=c++03 --inconclusive --enable=all --force ogrcartotablelayer.cpp
[ogrcartotablelayer.cpp:822]: (style) Variable 'bMustComma' is assigned a value that is never used.


git-svn-id: https://svn.osgeo.org/gdal/trunk@35292 f0d54148-0727-0410-94bb-9a71ac55c965