A very rough first take at a clang-format config file for GDAL C++ code.
DO NOT BLINDLY USE THIS VERSION

This file needs discussion and iteration before it should be used.

I have been trying to figure out what is the dominate usage for most
formatting questions in GDAL.  Occasionally, I've gone for a slightly
less common usage where I think there is a strong argument for clarity.

Overall, this is definitely not the style that I would pick on my own.
If I were starting from scratch, I would go with the Google style, but
that is where I work and spend most of my coding time.

git-svn-id: https://svn.osgeo.org/gdal/trunk@36467 f0d54148-0727-0410-94bb-9a71ac55c965